pyproject.toml
[project]
name = "agentup"
version = "0.1.0"
description = "CLI tool for creating A2A-compliant AI agents with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Luke Hinds", email = "luke@rdrocket.com" }
]
dependencies = [
    "click>=8.1.0",
    "questionary>=2.0.1",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.0",
    "httpx>=0.28.1",
    "a2a-sdk>=0.2.7",
    "uvicorn>=0.34.3",
    "fastapi>=0.115.12",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "numpy>=1.26.4",
    "pillow>=11.2.1",
    "asyncio>=3.4.3",
    "fastmcp>=2.8.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "mkdocs-material>=9.6.14",
    "authlib>=1.6.0",
    "psutil>=7.0.0",
]

[project.scripts]
agentup = "agent.cli.main:cli"

[project.optional-dependencies]
agent = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mkdocs-material>=9.6.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.uv]
package = true
